Update TBD and TSD adapters
