--- TAS_TBDAdapter/CRUD/Delete.cs
@@ -1,80 +1,0 @@
-/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2018, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *                                           
- *                                                                              
- * The BHoM is free software: you can redistribute it and/or modify         
- * it under the terms of the GNU Lesser General Public License as published by  
- * the Free Software Foundation, either version 3.0 of the License, or          
- * (at your option) any later version.                                          
- *                                                                              
- * The BHoM is distributed in the hope that it will be useful,              
- * but WITHOUT ANY WARRANTY; without even the implied warranty of               
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
- *                                                                            
- * You should have received a copy of the GNU Lesser General Public License     
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
- */
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using BH.oM.Base;
-using BHE = BH.oM.Environment;
-using BHG = BH.oM.Geometry;
-using System.Runtime.InteropServices;
-using BH.Engine.Environment;
-
-namespace BH.Adapter.TAS
-{
-    public partial class TasTBDAdapter : BHoMAdapter
-    {
-        /***************************************************/
-        /**** Public Methods                            ****/
-        /***************************************************/
-        /*
-        private bool Delete(BHE.Elements.Space space, IEnumerable<BHE.Elements.Space> spaces)
-        {
-
-            TBD.zone tbdZone = m_TBDDocument.Building.AddZone();
-            TBD.room tbdRoom = tbdZone.AddRoom();
-            tbdZone = Engine.TAS.Convert.ToTas(space, tbdZone);
-
-
-            foreach (BHE.Elements.BuildingElement element in space.BuildingElements)
-            {
-                //We have to add a building element to the zonesurface before we save the file. Otherwise we end up with a corrupt file!
-                TBD.buildingElement be = m_TBDDocument.Building.AddBuildingElement();
-
-                //Add zoneSrf and convert it
-                TBD.zoneSurface tbdZoneSrf = tbdZone.AddSurface();
-                tbdZoneSrf = Engine.TAS.Convert.ToTas(element.BuildingElementGeometry, tbdZoneSrf);
-
-                //Add roomSrf, create face, get its controlpoints and convert to TAS
-                TBD.Polygon tbdPolygon = tbdRoom.AddSurface().CreatePerimeter().CreateFace();
-                tbdPolygon = Engine.TAS.Convert.ToTas(element.BuildingElementGeometry.ICurve(), tbdPolygon);
-
-                //Set the building Element
-                tbdZoneSrf.buildingElement = Engine.TAS.Convert.ToTas(element, be);
-
-                //tasZoneSrf.type = BH.Engine.TAS.Query.GetSurfaceType(element, spaces);
-                tbdZoneSrf.orientation = (float)BH.Engine.Environment.Query.Azimuth(element.BuildingElementGeometry, new BHG.Vector());
-                //tasZoneSrf.orientation = BH.Engine.TAS.Query.GetOrientation(element.BuildingElementGeometry, bHoMSpace);
-                tbdZoneSrf.inclination = (float)BH.Engine.Environment.Query.Tilt(element.BuildingElementGeometry);
-                //tasZoneSrf.inclination = BH.Engine.TAS.Query.GetInclination(element.BuildingElementGeometry, bHoMSpace);
-            }
-
-            return true;
-        }
-        */
-
-
-        /***************************************************/
-
-
-    }
-}
