+++ TAS_TBDAdapter/TAS_TBDAdapter.csproj
@@ -32,6 +32,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Adapter_oM">
+      <HintPath>..\..\BHoM_Adapter\Build\Adapter_oM.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
     <Reference Include="Analytical_oM, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\..\BHoM\Build\Analytical_oM.dll</HintPath>
@@ -134,10 +138,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AdapterActions\Pull.cs" />
+    <Compile Include="AdapterActions\Push.cs" />
     <Compile Include="CRUD\Create.cs" />
-    <Compile Include="CRUD\Delete.cs" />
     <Compile Include="CRUD\Read.cs" />
-    <Compile Include="CRUD\Update.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Types\DependencyTypes.cs" />
     <Compile Include="TASTBDAdapter.cs" />
